requirements.txt
django==1.11.4
mysqlclient==1.3.10
Pillow==4.3.0

# API
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
pygments==2.2.0
coreapi==2.3.1

# Heroku
gunicorn==19.7.1
whitenoise==3.3.0

# Testing
factory-boy==2.9.2
prospector==0.12.7
